requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
gunicorn==21.2.0
weasyprint==62.3
reportlab==4.2.5
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5


# Windows (cualquier 64/32) → psycopg v3 con wheel binario
psycopg[binary]==3.2.9; sys_platform == "win32"

# No-Windows (Docker/Linux) → psycopg2-binary
psycopg2-binary==2.9.9; sys_platform != "win32"
